update url and view structure in the news app
