+++ server/portal/apps/news/api/__init__.py
@@ -1,0 +1,1 @@
+
