+++ server/portal/apps/news/api/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from portal.apps.news.views import UserNewsView
+from portal.apps.news.api.views import UserNewsView
 
 app_name = 'news_api'
 urlpatterns = [
