+++ server/portal/urls.py
@@ -88,7 +88,7 @@ urlpatterns = [
     path('api/datafiles/', include('portal.apps.datafiles.urls', namespace='datafiles')),
     path('api/system-monitor/', include('portal.apps.system_monitor.urls', namespace='system_monitor')),
     path('api/notifications/', include('portal.apps.notifications.urls', namespace='notifications')),
-    path('api/news/', include('portal.apps.news.api_urls', namespace='news_api')),
+    path('api/news/', include('portal.apps.news.api.urls', namespace='news_api')),
     path('api/jupyter_mounts/', include('portal.apps.jupyter_mounts.api.urls', namespace='jupyter_mounts_api')),
     path('api/projects/', include('portal.apps.projects.urls', namespace='projects')),
     path('api/site-search/', include('portal.apps.site_search.api.urls', namespace='site_search_api')),
